requirements.txt
fastapi
uvicorn[standard]
pydantic
python-keycloak
pydantic[email]
sqlalchemy
pyinvoke
alembic
python-dotenv
psycopg2